minor support for passing data to client
This doesn't actually get the latest timestamp from the server side (and
hardcodes 0), but it at least doesn't throw an error, either.

We will likely need to pass the JSON for the calls we're showing, since
we use it for transcriptions, etc. (Alternatively, we can load that when
opening the transcription view, but we already get it for any messages
that get loaded after the initial page load, so that would be weird.) I
haven't done that here, but it should be marginally simple to just loop
over and pass to handleLogEntry.