allow multi-line input for transcriptions
This defaults to two(+0.5ish, thanks browsers) lines of text, and
automatically increases the size of the textarea as more text gets
entered.

It doesn't automatically set the "correct" height of the entry form
because that would require it to run a layout first, which seems
impractical at the moment. The height will correct as soon as any text
change is made if it should be higher.